@@ -13,7 +13,7 @@ The ``ip_lists`` method implements IP :py:obj:`block- <block_ip>` and
    [botdetection.ip_lists]
 
    pass_ip = [
-    '140.238.172.132', # IPv4 of check.searx.space
+    '167.235.158.251', # IPv4 of check.searx.space
     '192.168.0.0/16',  # IPv4 private network
     'fe80::/10'        # IPv6 linklocal
    ]
@@ -33,15 +33,15 @@ from ipaddress import (
     IPv6Address,
 )
 
-from searx.tools import config
+from . import config
 from ._helpers import logger
 
 logger = logger.getChild('ip_limit')
 
 SEARXNG_ORG = [
     # https://github.com/searxng/searxng/pull/2484#issuecomment-1576639195
-    '140.238.172.132',  # IPv4 check.searx.space
-    '2603:c022:0:4900::/56',  # IPv6 check.searx.space
+    '167.235.158.251',  # IPv4 check.searx.space
+    '2a01:04f8:1c1c:8fc2::/64',  # IPv6 check.searx.space
 ]
 """Passlist of IPs from the SearXNG organization, e.g. `check.searx.space`."""
 
