@@ -8,8 +8,7 @@ import time
 import threading
 import os
 import signal
-from typing import Dict, Union, List, Any, Tuple, Optional
-from typing_extensions import TypedDict, Literal
+from typing import Any, Dict, List, Literal, Optional, Tuple, TypedDict, Union
 
 import redis.exceptions
 
