@@ -31,9 +31,11 @@ languages_file_header = """\
 # -*- coding: utf-8 -*-
 '''List of SearXNG's locale codes.
 
-This file is generated automatically by::
+.. hint::
 
-   ./manage pyenv.cmd searxng_extra/update/update_engine_traits.py
+   Don't modify this file, this file is generated by::
+
+     ./manage data.traits
 '''
 
 sxng_locales = (
@@ -100,8 +102,8 @@ def fetch_traits_map():
 def filter_locales(traits_map: EngineTraitsMap):
     """Filter language & region tags by a threshold."""
 
-    min_eng_per_region = 11
-    min_eng_per_lang = 13
+    min_eng_per_region = 15
+    min_eng_per_lang = 20
 
     _ = {}
     for eng in traits_map.values():
