@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
 
-from searx import plugins
+from searx import (
+    plugins,
+    limiter,
+    botdetection,
+)
+
 from mock import Mock
 from tests import SearxTestCase
 
@@ -46,6 +51,8 @@ class SelfIPTest(SearxTestCase):
         plugin = plugins.load_and_initialize_plugin('searx.plugins.self_info', False, (None, {}))
         store = plugins.PluginStore()
         store.register(plugin)
+        cfg = limiter.get_cfg()
+        botdetection.init(cfg, None)
 
         self.assertTrue(len(store.plugins) == 1)
 
